pyproject.toml
[tool.poetry]
name = "lotte"
version = "0.1.0"
description = ""
authors = ["Jaesung Ryu <cwryu6252@naver.com>"]

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
pyodbc = "^4.0.34"
azure-keyvault-secrets = "^4.5.1"
azure-identity = "^1.10.0"

[tool.poetry.dev-dependencies]
black = "^22.6.0"
isort = "^5.10.1"
flake8 = "^5.0.4"
mypy = "^0.971"
pytest = "^7.1.2"
coverage = "^6.4.3"
pytest-cov = "^3.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
